+++ sphinxcontrib/confluencebuilder/translator/__init__.py
@@ -1,7 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-    :copyright: Copyright 2017 by the contributors (see AUTHORS file).
-    :license: BSD-2-Clause, see LICENSE for details.
-"""
-
-__import__('pkg_resources').declare_namespace(__name__)