+++ test/unit-tests/common/test_common.py
@@ -30,21 +30,59 @@ class TestConfluenceCommon(unittest.TestCase):
     def test_admonitions(self):
         self._assertExpectedWithOutput('admonitions')
 
-    def test_bad_values(self):
-        self._assertExpectedWithOutput('badvalues')
+    def test_attribution(self):
+        self._assertExpectedWithOutput('attribution')
 
-    def test_formatting(self):
-        self._assertExpectedWithOutput('formatting')
+    def test_bibliographic(self):
+        self._assertExpectedWithOutput('bibliographic')
 
-    def test_list(self):
-        self._assertExpectedWithOutput('list')
+    def test_block_quotes(self):
+        self._assertExpectedWithOutput('block-quotes')
 
-    def test_references(self):
-        self._assertExpectedWithOutput('references')
+    def test_citations(self):
+        self._assertExpectedWithOutput('citations')
+
+    def test_contents(self):
+        raise unittest.SkipTest('pending pull request #126')
+        self._assertExpectedWithOutput('contents')
+
+    def test_definition_lists(self):
+        self._assertExpectedWithOutput('definition-lists')
+
+    def test_deprecated(self):
+        self._assertExpectedWithOutput('deprecated')
+
+    def test_epigraph(self):
+        self._assertExpectedWithOutput('epigraph')
+
+    def test_footnotes(self):
+        self._assertExpectedWithOutput('footnotes')
 
     def test_glossary(self):
         self._assertExpectedWithOutput('glossary')
-        self._assertExpectedWithOutput('glossary2')
+        self._assertExpectedWithOutput('glossary-ref')
+
+    def test_lists(self):
+        self._assertExpectedWithOutput('lists')
+
+    def test_list_table(self):
+        self._assertExpectedWithOutput('list-table')
+
+    def test_markup(self):
+        self._assertExpectedWithOutput('markup')
+
+    def test_option_lists(self):
+        self._assertExpectedWithOutput('option-lists')
+
+    def test_production_list(self):
+        self._assertExpectedWithOutput('production-list')
+
+    def test_raw(self):
+        self._assertExpectedWithOutput('raw')
+
+    def test_references(self):
+        self._assertExpectedWithOutput('references')
+        self._assertExpectedWithOutput('references-ref')
 
     def test_registry(self):
         # validate builder's registration into Sphinx
@@ -53,5 +91,11 @@ class TestConfluenceCommon(unittest.TestCase):
         else:
             self.assertTrue(EXT_NAME in self.app._extensions.keys())
 
+    def test_sections(self):
+        self._assertExpectedWithOutput('sections')
+
     def test_table(self):
         self._assertExpectedWithOutput('tables')
+
+    def test_transitions(self):
+        self._assertExpectedWithOutput('transitions')